Separated BErc20 and BEther test files
